format moved scripts
